:hammer: Move desktop-specific theme previews to desktopMain source set
